refactor: split out `HtmlContext`
